데이터 관리 방법 모색중
json을 사용하기엔 지금 당장은 힘들고
PlayerPrefs 사용하는 방향으로

+++ Assets/Scripts/Player.cs
@@ -12,7 +12,7 @@ public class Player
     public string Nickname { get; private set; } = "NULL";
     private bool Gender { get; set; }
 
-    public Player(string nickname, bool gender, int Sociability, int LearningAbility, int EconomicAbility, int SpecAbility)
+    public Player(string nickname, bool gender, float Sociability, float LearningAbility, float EconomicAbility, float SpecAbility)
     {
         Nickname = nickname;
         Gender = gender;
